+++ src/strongbox/core.clj
@@ -152,6 +152,9 @@
    ;; the list of addons from the catalogue
    :db nil
 
+   ;; the list of addons from the user-catalogue
+   :user-catalogue nil
+
    :log-lines []
 
    ;; a map of paths whose location may vary according to the cwd and envvars.
@@ -729,6 +732,14 @@
 
 ;;
 
+(defn-spec write-user-catalogue! nil?
+  "writes the `new-user-catalogue` to disk, using the current state of the `:user-catalogue` by default"
+  ([]
+   (write-user-catalogue! (catalogue/new-catalogue (get-state :user-catalogue :addon-summary-list))))
+  ([new-user-catalogue :catalogue/catalogue]
+   (catalogue/write-catalogue new-user-catalogue (paths :user-catalogue-file))
+   nil))
+
 (defn-spec get-create-user-catalogue (s/or :ok :catalogue/catalogue, :missing+no-create nil?)
   "returns the contents of the user catalogue at `user-catalogue-path`, creating one if `create?` is true (default)."
   ([]
@@ -748,17 +759,28 @@
 (defn-spec add-user-addon-list! nil?
   "adds a list of addons to the user catalogue"
   [addon-summary-list :addon/summary-list]
-  (let [user-catalogue-path (paths :user-catalogue-file)
-        user-catalogue (get-create-user-catalogue)
+  (let [user-catalogue (get-state :user-catalogue)
         tmp-catalogue (catalogue/new-catalogue addon-summary-list)
         new-user-catalogue (catalogue/merge-catalogues user-catalogue tmp-catalogue)]
-    (catalogue/write-catalogue new-user-catalogue user-catalogue-path))
+    (swap! state assoc :user-catalogue new-user-catalogue))
   nil)
 
 (defn-spec add-user-addon! nil?
   "adds a single addon to the user catalogue"
   [addon-summary :addon/summary]
   (add-user-addon-list! [addon-summary]))
+
+(defn-spec remove-user-addon! nil?
+  "removes a single addon from the user-catalogue"
+  [addon-summary :addon/summary]
+  (let [idx #(select-keys % [:source :source-id])
+        addon-summary-idx (idx addon-summary)
+        user-catalogue (->> (get-state :user-catalogue :addon-summary-list)
+                            (remove (fn [row]
+                                      (= (idx row) addon-summary-idx))))
+        new-user-catalogue (catalogue/new-catalogue user-catalogue)]
+    (swap! state assoc :user-catalogue new-user-catalogue))
+  nil)
 
 ;; catalogue db handling
 
@@ -791,6 +813,15 @@
   (swap! state update-in [:search] merge (select-keys -search-state-template [:page :results :selected-results-list]))
   nil)
 
+(defn-spec db-load-user-catalogue nil?
+  "loads the user catalogue into state, but only if it hasn't already been loaded."
+  []
+  (when-not (get-state :user-catalogue)
+    (let [create-user-catalogue? false
+          path (paths :user-catalogue-file)]
+      (swap! state assoc :user-catalogue (get-create-user-catalogue path create-user-catalogue?))))
+  nil)
+
 (defn-spec load-current-catalogue (s/or :ok :catalogue/catalogue, :error nil?)
   "merges the currently selected catalogue with the user-catalogue and returns the definitive list of addons 
   available to install. Handles malformed catalogue data by re-downloading catalogue."
@@ -814,9 +845,7 @@
                              (when-not testing?
                                (emergency-catalogue catalogue-location)))
 
-          create-user-catalogue? false
-          user-catalogue-data (p :p2/db:catalogue:read-user-catalogue
-                                 (get-create-user-catalogue (paths :user-catalogue-file) create-user-catalogue?))
+          user-catalogue-data (get-state :user-catalogue)
           ;; 2021-06-30: merge order changed. catalogue data is now merged over the top of the user-catalogue.
           ;; this is because the user-catalogue may now contain addons from all hosts and is likely to be out of date.
           final-catalogue (p :p2/db:catalogue:merge-catalogues
@@ -1142,7 +1171,7 @@
   "generates a list of 'export-records' from the addon summaries in the user catalogue and writes them to the given `output-file`"
   [output-file ::sp/file]
   (let [output-file (-> output-file fs/absolute str (utils/replace-file-ext ".json"))
-        catalogue (get-create-user-catalogue)
+        catalogue (get-state :user-catalogue)
         export (export-catalogue-addon-list catalogue)]
     (utils/dump-json-file output-file export)
     (info "wrote:" output-file)
@@ -1251,7 +1280,10 @@
    ;; downloads the big long list of addon information stored on github
   (download-current-catalogue)
 
-   ;; load the contents of the catalogue into the database
+  ;; load the user-catalogue. `db-load-catalogue` will incorporate this if it's found.
+  (db-load-user-catalogue)
+
+   ;; load the contents of the selected catalogue and the user catalogue
   (p :p2/db (db-load-catalogue))
 
    ;; match installed addons to those in catalogue